Gem prawn-fill form is modified to use the new version of LibreOffice
the code was modified to obtain the source of the template.pdf predesigned forms that will be used to fill the fields with the same size and color specifications.

Likewise coordinates are modified since the text came out misplaced, was achieved as a result align with the field space.

Likewise I will add a configuration that can be used checkbox.